support docker-compose
