+++ test/plugin/runner-helper/src/main/java/org/apache/skywalking/plugin/test/helper/RunningType.java
