--- test/plugin/runner/src/main/java/org/apache/skywalking/plugin/test/helper/vo/DependencyComponent.java
@@ -1,93 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.skywalking.plugin.test.helper.vo;
-
-import java.util.List;
-
-public class DependencyComponent implements DockerServiceReader {
-    private String image;
-    private String host;
-    private String version;
-    private List<String> expose;
-    private List<String> volumes;
-    private List<String> entrypoint;
-    private List<String> environment;
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public String getHost() {
-        return host;
-    }
-
-    public void setHost(String host) {
-        this.host = host;
-    }
-
-    public List<String> getEnvironment() {
-        return environment;
-    }
-
-    public void setEnvironment(List<String> environment) {
-        this.environment = environment;
-    }
-
-    @Override public List<String> environment() {
-        return environment;
-    }
-
-    @Override public String image() {
-        return host;
-    }
-
-    public List<String> getExpose() {
-        return expose;
-    }
-
-    public void setExpose(List<String> expose) {
-        this.expose = expose;
-    }
-
-    public List<String> getEntrypoint() {
-        return entrypoint;
-    }
-
-    public void setEntrypoint(List<String> entrypoint) {
-        this.entrypoint = entrypoint;
-    }
-
-    public List<String> getVolumes() {
-        return volumes;
-    }
-
-    public void setVolumes(List<String> volumes) {
-        this.volumes = volumes;
-    }
-}
