--- test/plugin/runner/src/main/java/org/apache/skywalking/plugin/test/helper/vo/DockerCompose.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.skywalking.plugin.test.helper.vo;
-
-import java.util.HashMap;
-
-public class DockerCompose {
-    private String version;
-    private HashMap<String, DockerService> services;
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public HashMap<String, DockerService> getServices() {
-        return services;
-    }
-
-    public void setServices(HashMap<String, DockerService> services) {
-        this.services = services;
-    }
-}
