--- test/plugin/runner/src/main/java/org/apache/skywalking/plugin/test/helper/vo/DockerService.java
@@ -1,106 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.skywalking.plugin.test.helper.vo;
-
-import com.google.common.collect.Lists;
-
-import java.util.List;
-
-public class DockerService {
-    private String name;
-    private String version;
-    private String image;
-    private String host;
-    private List<String> expose = Lists.newArrayList();
-    @Deprecated private List<String> ports;
-    private List<String> environments = Lists.newArrayList();
-    private List<String> volumes = Lists.newArrayList();
-
-    @Deprecated private List<String> links;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getVersion() {
-        return version;
-    }
-
-    public void setVersion(String version) {
-        this.version = version;
-    }
-
-    public String getImage() {
-        return image;
-    }
-
-    public void setImage(String image) {
-        this.image = image;
-    }
-
-    public List<String> getExpose() {
-        return expose;
-    }
-
-    public void setExpose(List<String> expose) {
-        this.expose = expose;
-    }
-
-    public List<String> getPorts() {
-        return ports;
-    }
-
-    public void setPorts(List<String> ports) {
-        this.ports = ports;
-    }
-
-    public List<String> getEnvironments() {
-        return environments;
-    }
-
-    public void setEnvironments(List<String> environments) {
-        this.environments = environments;
-    }
-
-    public List<String> getVolumes() {
-        return volumes;
-    }
-
-    public void setVolumes(List<String> volumes) {
-        this.volumes = volumes;
-    }
-
-    public List<String> getLinks() {
-        return links;
-    }
-
-    public void setLinks(List<String> links) {
-        this.links = links;
-    }
-
-    public String getHost() {
-        return host;
-    }
-
-    public void setHost(String host) {
-        this.host = host;
-    }
-}
